+++ pycons3rt3/cons3rtdefaults.py
@@ -6,7 +6,7 @@ A module for default values.
 """
 
 # Default number of CPUs to include when subscripting to a template
-default_template_subscription_max_cpus = 32
+default_template_subscription_max_cpus = 20
 
 # Default number RAM in MBs to include when subscripting to a template
 default_template_subscription_max_ram_mb = 131072